+++ app/scripts/AddTrackDialog.js
@@ -1,9 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import Button from './Button';
 import Dialog from './Dialog';
 import TilesetFinder from './TilesetFinder';
+import TrackSourceEditor from './TrackSourceEditor';
 import PlotTypeChooser from './PlotTypeChooser';
+
+import withModal from './hocs/with-modal';
 
 // Configs
 import { AVAILABLE_TRACK_TYPES } from './configs';
@@ -14,7 +18,7 @@ import {
 } from './utils';
 
 // Styles
-import '../styles/AddTrackDialog.module.scss';
+import styles from '../styles/AddTrackDialog.module.scss';
 
 class AddTrackDialog extends React.Component {
   constructor(props) {
@@ -25,12 +29,42 @@ class AddTrackDialog extends React.Component {
     this.state = {
       selectedTilesets: [{ datatype: 'none' }],
       allTracksSameDatatype: true, // Do all selected tracks have the same datatype
+      activeTab: this.getActiveTab(),
     };
 
+    this.handleNextBound = this.handleNext.bind(this);
+    this.handlePlotTypeSelectedBound = this.handlePlotTypeSelected.bind(this);
+    this.handleSearchBoxBound = this.handleSearchBox.bind(this);
+    this.handleSubmitAndCloseBound = this.handleSubmitAndClose.bind(this);
     this.handleSubmitBound = this.handleSubmit.bind(this);
 
     this.handleTilesetPickerDoubleClickBound = this.handleTilesetPickerDoubleClick.bind(this);
+    this.handleTrackChosenBound = this.handleTrackChosen.bind(this);
+    this.handleTrackSourceSavedBound = this.handleTrackSourceSaved.bind(this);
     this.selectedTilesetsChangedBound = this.selectedTilesetsChanged.bind(this);
+  }
+
+  componentWillUpdate(nextProps, nextState) {
+    if (nextState.selectedTilesets.length === 0) {
+      this.selectedPlotType = undefined;
+    }
+  }
+
+  componentDidUpdate(props, state) {
+    if (this.updateTab) {
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState({
+        activeTab: this.getActiveTab()
+      });
+      this.updateTab = false;
+    }
+  }
+
+  getActiveTab() {
+    if (this.props.trackSourceServers.length === 0) return 'trackSources';
+    if (!this.state || this.state.selectedTilesets.length === 0) return 'datasets';
+    if (!this.selectedPlotType && !this.props.hidePlotTypeChooser) return 'trackTypes';
+    return 'done';
   }
 
   /**
@@ -61,14 +95,29 @@ class AddTrackDialog extends React.Component {
     return orientation;
   }
 
+  handleNext() {
+    this.setState({
+      activeTab: this.getActiveTab()
+    });
+  }
+
   handleSubmit(evt) {
     if (evt) evt.preventDefault();
+
+    this.state.selectedTilesets.forEach((tileset) => {
+      this.addDefaultTrackType(tileset);
+    });
 
     this.props.onTracksChosen(
       this.state.selectedTilesets,
       this.props.position,
       this.props.host
     );
+  }
+
+  handleSubmitAndClose(evt) {
+    this.handleSubmit(evt);
+    this.props.modal.close();
   }
 
   /**
@@ -80,12 +129,30 @@ class AddTrackDialog extends React.Component {
    *  tileset: { uuid: 'CXCX', filetype: 'cooler' ....}
    */
   handleTilesetPickerDoubleClick(tileset) {
-    this.selectedTilesetsChanged([tileset]);
+    this.addDefaultTrackType(tileset);
+
+    // This is an anti-pattern. We update the state of `selectedTilesets` but
+    // we don't until until it's changed and instead use it directly. To avoid
+    // inconsistencies I am returning the new state value to be able to use it
+    // here directly.
+    const selectedTilesets = this.selectedTilesetsChanged([tileset]);
 
     // should iterate over the selected tilesets
-    this.props.onTracksChosen(this.state.selectedTilesets,
+    this.props.onTracksChosen(
+      selectedTilesets,
       this.props.position,
-      this.props.host);
+      this.props.host
+    );
+  }
+
+  addDefaultTrackType(tileset) {
+    if (!tileset.type) {
+      // Let's use the first of the available track types
+      tileset.type = AVAILABLE_TRACK_TYPES(
+        this.datatypes(),
+        this.getOrientation(this.props.position)
+      )[0].type;
+    }
   }
 
   handleOptionsChanged(newOptions) {
@@ -99,9 +166,24 @@ class AddTrackDialog extends React.Component {
 
     this.selectedPlotType = newPlotType;
 
+    this.updateTab = true;
+
     this.setState({
       selectedTilesets,
     });
+  }
+
+  handleSearchBox(element) {
+    this.tilesetFinderSearchBox = element;
+  }
+
+  handleTrackChosen(value) {
+    this.props.onTracksChosen(value, this.props.position);
+  }
+
+  async handleTrackSourceSaved(sources) {
+    await this.props.onTrackSourceChanged(sources);
+    this.setState({ activeTab: this.getActiveTab() });
   }
 
   selectedTilesetsChanged(selectedTilesetsIn) {
@@ -148,54 +230,264 @@ class AddTrackDialog extends React.Component {
     }
 
     this.setState({ selectedTilesets, allTracksSameDatatype });
+
+    return selectedTilesets;
+  }
+
+  open(activeTab) {
+    return () => this.setState({ activeTab });
+  }
+
+  datatypes() {
+    return this.state.selectedTilesets.map((x) => {
+      if (x.filetype === 'cooler') {
+        // cooler files can also supply chromsizes
+        return [x.datatype, 'chromsizes'];
+      }
+
+      return [x.datatype];
+    });
   }
 
   render() {
     const orientation = this.getOrientation(this.props.position);
-    const form = (
-      <div>
-        <TilesetFinder
-          // Only for testing purposes
-          ref={(c) => { this.tilesetFinder = c; }}
-          datatype={this.props.datatype}
-          onDoubleClick={this.handleTilesetPickerDoubleClick.bind(this)}
-          onTracksChosen={value => this.props.onTracksChosen(value, this.props.position)}
-          orientation={orientation}
-          selectedTilesetChanged={this.selectedTilesetsChanged.bind(this)}
-          trackSourceServers={this.props.trackSourceServers}
-        />
-      </div>
-    );
 
     return (
       <Dialog
-        maxHeight={true}
+        maxHeight={46.75}
         okayTitle="Submit"
         onCancel={this.props.onCancel}
         onOkay={this.handleSubmitBound}
         title="Add Track"
       >
-        { form }
+        <div
+          className={
+            this.state.activeTab === 'trackSources'
+              ? styles.addTrackDialogToggableOpen
+              : styles.addTrackDialogToggable
+          }
+        >
+          <div
+            className={
+              this.state.activeTab === 'trackSources'
+                ? styles.addTrackDialogTogglerActive
+                : styles.addTrackDialogToggler
+            }
+          >
+            <button onClick={this.open('trackSources')} type="button">
+              {this.props.trackSourceServers.length >= 1 && (
+                <span className={styles.addTrackDialogTogglerLabel}>
+                  Change track source server(s):
+                </span>
+              )}
+              {this.props.trackSourceServers.length === 0 && (
+                <span className={styles.addTrackDialogTogglerLabel}>
+                  Set track source servers
+                </span>
+              )}
+              <span className={styles.addTrackDialogTogglerValue}>
+              {this.props.trackSourceServers.length > 0 && (
+                <span>{this.props.trackSourceServers[0]}</span>
+              )}
+              </span>
+              <span className={styles.addTrackDialogTogglerTriangle} />
+            </button>
+          </div>
+          <div
+            className={
+              this.state.activeTab === 'trackSources'
+                ? styles.addTrackDialogToggableContentOpen
+                : styles.addTrackDialogToggableContent
+            }
+          >
+            {this.state.activeTab === 'trackSources' && (
+              <TrackSourceEditor
+                onTrackSourceChanged={this.props.onTrackSourceChanged}
+                onTrackSourceSaved={this.handleTrackSourceSavedBound}
+                trackSources={this.props.trackSourceServers}
+              />
+            )}
+          </div>
+        </div>
+        <div
+          className={
+            this.state.activeTab === 'datasets'
+              ? styles.addTrackDialogToggableOpen
+              : styles.addTrackDialogToggable
+          }
+        >
+          <div
+            className={
+              this.state.activeTab === 'datasets'
+                ? styles.addTrackDialogTogglerActive
+                : styles.addTrackDialogToggler
+            }
+          >
+            <button onClick={this.open('datasets')} type="button">
+              {this.state.selectedTilesets.length >= 1 && (
+                <span className={styles.addTrackDialogTogglerLabel}>
+                  Change dataset(s):
+                </span>
+              )}
+              {this.state.selectedTilesets.length === 0 && (
+                <span className={styles.addTrackDialogTogglerLabel}>
+                  Select datasets
+                </span>
+              )}
+              <span className={styles.addTrackDialogTogglerValue}>
+              {this.state.selectedTilesets.length > 0 && (
+                <span>{this.state.selectedTilesets[0].name}</span>
+              )}
+              </span>
+              <span className={styles.addTrackDialogTogglerTriangle} />
+            </button>
+          </div>
+          <div
+            className={
+              this.state.activeTab === 'datasets'
+                ? styles.addTrackDialogToggableContentOpen
+                : styles.addTrackDialogToggableContent
+            }
+          >
+            {this.state.activeTab === 'datasets' && (
+              <div>
+                <TilesetFinder
+                  // Only for testing purposes
+                  ref={(c) => { this.tilesetFinder = c; }}
+                  datatype={this.props.datatype}
+                  onDoubleClick={this.handleTilesetPickerDoubleClickBound}
+                  onTracksChosen={this.handleTrackChosenBound}
+                  orientation={orientation}
+                  searchBox={this.handleSearchBoxBound}
+                  selectedTilesetChanged={this.selectedTilesetsChangedBound}
+                  selectedTilesets={this.state.selectedTilesets.map(x => x.serverUidKey)}
+                  trackSourceServers={this.props.trackSourceServers}
+                />
+                <Button
+                  className={styles.addTrackDialogNext}
+                  disabled={this.state.selectedTilesets.length === 0}
+                  onClick={this.handleNextBound}
+                >
+                  Done
+                </Button>
+              </div>
+            )}
+          </div>
+        </div>
+        {!this.props.hidePlotTypeChooser && (
+          <div
+            className={
+              this.state.activeTab === 'trackTypes'
+                ? styles.addTrackDialogToggableOpen
+                : styles.addTrackDialogToggable
+            }
+          >
+            <div
+              className={
+                this.state.activeTab === 'trackTypes'
+                  ? styles.addTrackDialogTogglerActive
+                  : styles.addTrackDialogToggler
+              }
+            >
+              <button
+                disabled={this.state.selectedTilesets.length === 0}
+                onClick={this.open('trackTypes')}
+                type="button"
+              >
+                {this.selectedPlotType ? (
+                  <span>
+                    <span>Change track type: </span>
+                    <span className={styles.addTrackDialogTogglerValue}>
+                      {this.selectedPlotType}
+                    </span>
+                  </span>
+                ) : (
+                  <span>Select track type</span>
+                )}
+                <span className={styles.addTrackDialogTogglerTriangle} />
+              </button>
+            </div>
+            <div
+              className={
+                this.state.activeTab === 'trackTypes'
+                  ? styles.addTrackDialogToggableContentOpen
+                  : styles.addTrackDialogToggableContent
+              }
+            >
+              {this.state.activeTab === 'trackTypes' && (
+                <PlotTypeChooser
+                  // Only for testing purposes
+                  ref={(c) => { this.plotTypeChooser = c; }}
+                  allTracksSameDatatype={this.state.allTracksSameDatatype}
+                  datatypes={this.datatypes()}
+                  onPlotTypeSelected={this.handlePlotTypeSelectedBound}
+                  orientation={orientation}
+                  // plotType={this.selectedPlotType}
+                  position={this.props.position}
+                />
+              )}
+            </div>
+          </div>
+        )}
         {
-          (
-            <PlotTypeChooser
-              // Only for testing purposes
-              ref={(c) => { this.plotTypeChooser = c; }}
-              allTracksSameDatatype={this.state.allTracksSameDatatype}
-              datatypes={this.state.selectedTilesets.map((x) => {
-                if (x.filetype === 'cooler') {
-                  // cooler files can also supply chromsizes
-                  return [x.datatype, 'chromsizes'];
-                }
-
-                return [x.datatype];
-              })}
-              onPlotTypeSelected={this.handlePlotTypeSelected.bind(this)}
-              orientation={orientation}
-              position={this.props.position}
-            />
-          )
+          // <div
+          //   styleName={
+          //     this.state.activeTab === 'configurations'
+          //       ? 'add-track-dialog-toggable-open'
+          //       : 'add-track-dialog-toggable'
+          //   }
+          // >
+          //   <div
+          //     styleName={
+          //       this.state.activeTab === 'configurations'
+          //         ? 'add-track-dialog-toggler-active'
+          //         : 'add-track-dialog-toggler'
+          //     }
+          //   >
+          //     <button
+          //       disabled={
+          //         this.state.selectedTilesets.length === 0 || !this.selectedPlotType
+          //       }
+          //       onClick={this.open('configurations')}
+          //       type="button"
+          //     >
+          //       <span>Configure track</span>
+          //       <span className=styles.addTrackDialogTogglerTriangle} />
+          //     </button>
+          //   </div>
+          //   <div
+          //     className={
+          //       this.state.activeTab === 'configurations'
+          //         ? styles.addTrackDialogToggableContentOpen
+          //         : styles.addTrackDialogToggableContent
+          //     }
+          //   >
+          //     {this.state.activeTab === 'configurations' && (
+          //       <p>Nice!</p>
+          //     )}
+          //   </div>
+          //   <div
+          //     className={
+          //       this.state.activeTab === 'done'
+          //         ? styles.addTrackDialogToggableContentOpen
+          //         : styles.addTrackDialogToggableContent
+          //     }
+          //   >
+          //     {this.state.activeTab === 'done' && (
+          //       <Button>Add Track</Button>
+          //     )}
+          //   </div>
+          // </div>
         }
+        {this.state.activeTab === 'done' && (
+          <Button
+            className={styles.addTrackDialogSubmit}
+            onClick={this.handleSubmitAndCloseBound}
+            primary={true}
+          >
+            {this.state.selectedTilesets.length > 1 ? 'Add Tracks!' : 'Add Track!'}
+          </Button>
+        )}
       </Dialog>
     );
   }
@@ -208,10 +500,12 @@ AddTrackDialog.defaultProps = {
 AddTrackDialog.propTypes = {
   datatype: PropTypes.string.isRequired,
   host: PropTypes.string.isRequired,
+  modal: PropTypes.object.isRequired,
   onCancel: PropTypes.func.isRequired,
   onTracksChosen: PropTypes.func.isRequired,
+  onTrackSourceChanged: PropTypes.func.isRequired,
   position: PropTypes.string,
   trackSourceServers: PropTypes.array.isRequired,
 };
 
-export default AddTrackDialog;
+export default withModal(AddTrackDialog);