+++ app/scripts/Modal.js
@@ -7,6 +7,16 @@ import withModal from './hocs/with-modal';
 
 import '../styles/Modal.module.scss';
 
+const getRowHeight = (props) => {
+  if (props.fullHeight) {
+    return '4rem auto 5rem';
+  }
+  if (Number.isFinite(props.maxHeight)) {
+    return `4rem minmax(auto, ${props.maxHeight - 9}rem) 5rem`;
+  }
+  return null;
+};
+
 const Modal = (props) => {
   const handleClose = () => {
     props.modal.close();
@@ -16,11 +26,21 @@ const Modal = (props) => {
   return (
     <div styleName={`modal-background ${props.hide ? 'modal-hide' : ''}`}>
       <div styleName="modal-wrap">
-        <div styleName={`modal-window ${props.maxHeight ? 'modal-window-max-height' : ''}`}>
+        <div
+          style={{
+            maxHeight: Number.isFinite(props.maxHeight) ? `${props.maxHeight}rem` : ''
+          }}
+          styleName={`modal-window ${props.fullHeight || Number.isFinite(props.maxHeight) ? 'modal-window-full-height' : ''}`}
+        >
           {props.closeButton && (
             <Button onClick={handleClose}><Cross /></Button>
           )}
-          <div styleName="modal-content">
+          <div
+            style={{
+              gridTemplateRows: getRowHeight(props)
+            }}
+            styleName={`modal-content ${props.fullHeight ? 'modal-content-full-height' : ''} ${props.maxHeight ? 'modal-content-max-height' : ''}`}
+          >
             {props.children}
           </div>
         </div>
@@ -32,14 +52,16 @@ const Modal = (props) => {
 Modal.defaultProps = {
   closeButton: true,
   hide: false,
-  maxHeight: false,
+  fullHeight: false,
+  maxHeight: Number.NaN,
 };
 
 Modal.propTypes = {
   children: PropTypes.element.isRequired,
   closeButton: PropTypes.bool,
   hide: PropTypes.bool,
-  maxHeight: PropTypes.bool,
+  fullHeight: PropTypes.bool,
+  maxHeight: PropTypes.number,
   modal: PropTypes.object.isRequired,
   onClose: PropTypes.func
 };