+++ app/scripts/TiledPlot.js
@@ -288,6 +288,7 @@ class TiledPlot extends React.Component {
           host={this.state.addTrackHost}
           onCancel={this.handleNoTrackAddedBound}
           onTracksChosen={this.handleTracksAddedBound}
+          onTrackSourceChanged={this.props.onTrackSourceChanged}
           position={this.state.addTrackPosition || this.props.addTrackPosition}
           trackSourceServers={this.props.trackSourceServers}
         />
@@ -307,6 +308,7 @@ class TiledPlot extends React.Component {
           onTracksChosen={(newTrack) => {
             this.handleDivisorChosen(series, newTrack);
           }}
+          onTrackSourceChanged={this.props.onTrackSourceChanged}
           trackSourceServers={this.props.trackSourceServers}
         />
       );
@@ -2299,6 +2301,7 @@ TiledPlot.propTypes = {
   onScalesChanged: PropTypes.func,
   onTrackOptionsChanged: PropTypes.func,
   onTrackPositionChosen: PropTypes.func,
+  onTrackSourceChanged: PropTypes.func,
   onTracksAdded: PropTypes.func,
   onUnlockValueScale: PropTypes.func,
   onValueScaleChanged: PropTypes.func,